+++ packages/@baldeweg/auth/src/i18n.js
@@ -3,7 +3,7 @@ import en_US from './../locales/en-US.json'
 import de_DE from './../locales/de-DE.json'
 
 export default createI18n({
-  locale: import.meta.env.VUE_APP_I18N_LOCALE,
+  locale: 'en-US',
   fallbackLocale: import.meta.env.VUE_APP_I18N_FALLBACK_LOCALE || 'en-US',
   messages: {
     'en-US': en_US,