+++ packages/@baldeweg/web-demo/src/views/HomeView.vue
@@ -1,5 +1,6 @@
 <script setup>
 import { useTitle } from '@baldeweg/ui/composables/useTitle'
+import i18n from '@/i18n.js'
 
 useTitle({
   title: 'Home - Demo',
@@ -8,4 +9,7 @@ useTitle({
 
 <template>
   <h1>{{ $t('welcome') }}</h1>
+
+  Available locales: {{ i18n.global.availableLocales }}<br />
+  Selected locale: {{ i18n.global.locale}}<br />
 </template>